Moved Region to static reference Set height to ushort

+++ Client/Game/Enviroment/BiomeManagement/Biome.cs
@@ -7,8 +7,8 @@ namespace VoxelValley.Client.Game.Enviroment.BiomeManagement
         public abstract string Name { get; }
         public abstract Color Color { get; }
         public abstract byte BiomeId { get; } //TODO: Namespaced
-        public abstract short GetHeight(int x, int z);
-        internal abstract ushort GetVoxel(int x, int y, int z, int height);
-        internal abstract void GetFinishers(int worldX, int worldZ, int height, ref ushort[] voxels);
+        public abstract ushort GetHeight(int x, int z);
+        internal abstract ushort GetVoxel(int x, int y, int z, ushort height);
+        internal abstract void GetFinishers(int worldX, int worldZ, ushort height, ref ushort[] voxels);
     }
 }
