+++ Client/Game/Enviroment/BiomeManagement/Biomes/Desert/Dunes.cs
@@ -10,12 +10,12 @@ namespace VoxelValley.Client.Game.Enviroment.BiomeManagement.Biomes.Desert
         public override Color Color { get => Color.Orange; }
         public override byte BiomeId { get => 51; }
 
-        public override short GetHeight(int x, int z)
+        public override ushort GetHeight(int x, int z)
         {
             return GenerationUtilities.MapToWorldByte(GenerationUtilities.FBMPerlin(x, z, 5, 0.4f, 0.25f));
         }
 
-        internal override ushort GetVoxel(int x, int y, int z, int height)
+        internal override ushort GetVoxel(int x, int y, int z, ushort height)
         {
 
             if (y > height)
@@ -28,7 +28,7 @@ namespace VoxelValley.Client.Game.Enviroment.BiomeManagement.Biomes.Desert
                 return VoxelManager.GetVoxel("stone").Id;
         }
 
-        internal override void GetFinishers(int worldX, int worldZ, int height, ref ushort[] voxels)
+        internal override void GetFinishers(int worldX, int worldZ, ushort height, ref ushort[] voxels)
         {
 
         }
